third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/settings/shortcuts-a11y-test-expected.txt

Tests accessibility in the settings tool shortcuts pane using the axe-core linter.

Running: testShortcuts
aXe violations: []
